requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer Vision & Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Scientific Computing
numpy>=1.24.0

# Machine Learning & Evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Progress Bars
tqdm>=4.65.0

# Experiment Tracking (Optional but recommended)
tensorboard>=2.13.0
# wandb>=0.15.0  # Uncomment if using Weights & Biases

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration Management
pyyaml>=6.0.0
